Separate TimeZone into separate component
